+++ src/main/java/subway/domain/vo/Money.java
@@ -1,4 +1,4 @@
-package subway.domain;
+package subway.domain.vo;
 
 import java.util.Objects;
