+++ src/main/java/subway/domain/vo/Name.java
@@ -1,22 +1,22 @@
-package subway.domain;
+package subway.domain.vo;
 
 import java.util.Objects;
 
-public class Color {
+public class Name {
 
     private final String value;
 
-    public Color(final String value) {
+    public Name(final String value) {
         validate(value);
         this.value = value;
     }
 
     private void validate(final String value) {
         if (Objects.isNull(value) || value.isBlank()) {
-            throw new IllegalArgumentException("색상은 null 또는 공백일 수 없습니다.");
+            throw new IllegalArgumentException("이름은 null 또는 공백일 수 없습니다.");
         }
         if (value.length() > 10) {
-            throw new IllegalArgumentException("색상은 1 ~ 10글자여야합니다.");
+            throw new IllegalArgumentException("이름은 1 ~ 10글자여야합니다.");
         }
     }
 
@@ -28,8 +28,8 @@ public class Color {
     public boolean equals(final Object o) {
         if (this == o) return true;
         if (o == null || getClass() != o.getClass()) return false;
-        final Color color = (Color) o;
-        return Objects.equals(value, color.value);
+        final Name name = (Name) o;
+        return Objects.equals(value, name.value);
     }
 
     @Override
@@ -39,7 +39,7 @@ public class Color {
 
     @Override
     public String toString() {
-        return "Color{" +
+        return "Name{" +
                 "value='" + value + '\'' +
                 '}';
     }
