+++ src/main/java/subway/repository/SectionRepository.java
@@ -5,7 +5,7 @@ import subway.dao.entity.SectionEntity;
 import subway.dao.entity.StationEntity;
 import subway.dao.SectionDao;
 import subway.dao.StationDao;
-import subway.domain.Distance;
+import subway.domain.vo.Distance;
 import subway.domain.Section;
 import subway.domain.Station;
