--- src/test/java/subway/domain/SubwayTest.java
@@ -1,15 +1,0 @@
-package subway.domain;
-
-import org.junit.jupiter.api.DisplayNameGeneration;
-import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.DisplayNameGenerator.ReplaceUnderscores;
-
-@DisplayNameGeneration(ReplaceUnderscores.class)
-class SubwayTest {
-
-    @Test
-    void 모든_구간을_가져온다() {
-
-    }
-}
