+++ src/test/java/subway/fixture/SectionFixture.java
@@ -1,6 +1,6 @@
 package subway.fixture;
 
-import subway.domain.Distance;
+import subway.domain.vo.Distance;
 import subway.domain.Section;
 import subway.domain.Station;
